ref(flags): add settings button to issues CTA
