+++ static/app/components/events/featureFlags/featureFlagInlineCTA.tsx
@@ -4,6 +4,7 @@ import {usePrompt} from 'sentry/actionCreators/prompts';
 import ButtonBar from 'sentry/components/buttonBar';
 import {Button, LinkButton} from 'sentry/components/core/button';
 import {DropdownMenu} from 'sentry/components/dropdownMenu';
+import FeatureFlagSettingsButton from 'sentry/components/events/featureFlags/featureFlagSettingsButton';
 import {useFeatureFlagOnboarding} from 'sentry/components/events/featureFlags/useFeatureFlagOnboarding';
 import {IconClose, IconMegaphone} from 'sentry/icons';
 import {t} from 'sentry/locale';
@@ -55,7 +56,12 @@ export default function FeatureFlagInlineCTA({projectId}: {projectId: string}) {
     return null;
   }
 
-  const actions = <ButtonBar gap={1}>{feedbackButton}</ButtonBar>;
+  const actions = (
+    <ButtonBar gap={1}>
+      {feedbackButton}
+      <FeatureFlagSettingsButton orgSlug={organization.slug} />
+    </ButtonBar>
+  );
 
   return (
     <InterimSection
